+++ internal/app/collection_find_queries.go
@@ -3,7 +3,6 @@ package app
 import (
 	"encoding/json"
 	"io/ioutil"
-	"os"
 	"path"
 
 	"github.com/wailsapp/wails/v2/pkg/runtime"
@@ -25,7 +24,7 @@ func updateQueryFile(a *App, newData map[string]SavedQuery) error {
 		return err
 	}
 
-	err = ioutil.WriteFile(filePath, jsonData, os.ModePerm)
+	err = ioutil.WriteFile(filePath, jsonData, 0644)
 	return err
 }