+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Patches:
 * Preserve state after switching to another tab (#56).
 * Find view: ask for confirmation before negligently deleting documents when the user has clicked the '-' button (#58).
 * Set a deadline for counting documents, and added a button to count documents if the deadline has been exceeded.
+* Changed os.ModePerm (777) file permissions to 644 and 755.
 * UI improvements.
 
 Bugfixes: